Replace the edit and delete icons with a single dropdown menu button that allows for editing, deleting and viewing details.
Gift details:  Can now see a gift's details: Who added the item and
when.  Who the gift is for, in case it's for more than one person. 
And if the gift is a surprise we say that in bold.

Split up giftlist.html into smaller partial pages: addgiftform.html,
buygiftform.html, editgiftform.html and viewitem.html