Doing away with "Past Events" section (It's still there in this commit but will be gone in the next.)
The problem with the "Past Events" section is that it would also have
to contain the monstrous code that I use to edit events.  So there's
ddbtn-addcircle.html, mycircles.html and myexpiredcircles.html that
all have very similar code for adding/editing events.

So to keep from repeating all this code a third time, I am doing away
with the separate "Past Events" section.

Now, past events are in the same section as active events.

See User.circleList.  That method used to return just the active
circles.  Now it returns active and past events.  It splits a user's
circles into 2 groups - active and expired.  It sorts the active
events by date with the ascending, so the nearest event is at the top.

The expired events are sorted by date descending so the most recently
expired events are at the top.

The 2 list of circles are then added back to a single list of circles
that is displayed in the "Events" section