THE BIG MOD HERE IS THE IMPLEMENTATION OF REMINDERS
Boot:  Added Reminder.boot to read the reminder table and create the
executors

RestService: insertReminders, deleteReminders

Reminder: moved a lot of the code from the companion object to the
class.  Just seemed to make more sense to say r.createReminderExecutor
than Reminder.createReminderExecutor(r)

lbb.css:  changed .shadow so it wouldn't be so harsh

app.js:  implemented Reminders

mycircles.html:  flip flopped the placement of 2 elements of the
edit-circle dialog:  There's the area where you select the people to
add, and there's the section where you display who is currently in the
list.  Used to be that the current participants were listed on the
left and the field to add people was on the right.  Now, the current
participants are listed on the right and the field to add people is on
the left - just looked a little better

reminders.html:  finished with the first pass - does everything I need
it to do