Change the way reminders are deleted when a circle is deleted.  Now there is a new method on DateChangeListener: dateDeletedSet that fires when date_deleted.apply() is called - THIS is where the reminders are deleted.  Otherwise, we would have to remember to delete reminders whereever we may set date_deleted.
That's what we did in the prior version of RestService.  In
RestService.updateCircle, there is the case that detects
"date_deleted" being set.  Just under the call to
circle.date_deleted(...), we had a call to
circle.reminders.foreach(_.delete_!).

Now that we have the new method in DateChangeListener, we can call
circle.date_deleted anywhere we want and not have to remember to also
delete the reminders.