Changed the way we show the details of a gift.  No more "Details" item from the caret/dropdown.  Instead, hover over the description to get a twitter bootstrap popover.  See GiftCtrl.popoverOptions and viewitem.html.
Added date.js to the project to handle the "pretty" date formatting
you see in the popover.

Also, show edit and delete links even when you are not allowed to
edit/delete.  We pop an alert telling you you don't have perms if
that's the case.  The reason I did this is because, now that we don't
have the "Details" item anymore, there was the potential to have that
caret/dropdown but not items underneath - would definitely be
confusing.  And didn't want to have some items with the caret and
other without it.