FB login implementation: little by little
Don't know what's going on with fb.  Take foo.html - the fb login
button on that page works every time.  Copy foo.html over to
login.html and it only seems to work some of the time.

Thought it had something to do with the code in $scope.initfbuser()
(app-ConnectCtrl.js) because as soon as I would start uncommenting
code in that method, the fb login would start acting up again (the
dialog wouldn't dismiss and the "success" would stop being called).

The fix/hack:  So I would just re-copy (over and over again) foo.html,
overlaying it on top of login.html (but the 2 pages are identical
except for some "tagging" text in each page - "login.html" and
"foo.html" to let me know exactly which page I was looking at.)

At this point, I have $scope.initfbuser() working up to the point
where it will recognized that a fb user also has an LBB account and so
it will automatically send the person to his wishlist page.

Note the $routeChangeStart event that we listen for in app.js.  This
code fires on every request.  It is here that we determine if we let
the user through or redirect to /login.  One of the things we look for
is the presence of $rootScope.user.

$rootScope.user is new.  We used to just store $scope.user and $emit
"userchange" events.  But it sounds like it is better to keep shared
variables and functions in the $rootScope.