In the prior commit, I was logging in the fb - the 'person' record already had a facebook_id value, so this was a "merged" account.
In this commit, I am able to login to fb where the 'person' record
does not yet have a facebook_id value.  After logging in to fb, I
checked the person table and verified that it got updated with the
facebook_id.

This is an example of "merging" fb and lbb accounts because the record
in the person table got updated with a facebook id!