Switched the localhost app back to running on 8080.  80 always gives the error "address already in use" and I have to do that stupid netstat -a -o -n followed by taskkill /F /PID [pid]
Also, moved fbinvite from app-ConnectCtrl to app-FriendCtrl because
the Invite button is on friends.html, which used to reference
app-ConnectCtrl but doesn't any longer.

Finally, changed the signature of fbinvite so that it doesn't take a
'friend' argument.  Facebook already provides a nice dialog that lets
you choose which friends to invite.