Nice demonstration of deleting facebook request id's once they've been accepted.
See Gift2Ctrl.  We look at the url and if it contains facebook request
id's, we delete them.  The user is in facebook and see the app
request.  The user clicks the app request which sends the user to LBB.
 The LBB url contains facebook request id(s).  We parse those from the
url and delete them.  When you go back to facebook, you'll see they
are gone!

Note: It appears that the person sending the app requests cannot
delete them.  They have to be deleted during the course of accepting
the request BY THE PERSON WHO RECEIVED THE REQUEST.

So if Brent sends the app request to Tamie, it will be deleted when
Tamie logs in and accepts the request.