App requests working differently now.  Have a new app_request table in the db and a new AppRequest entity class.  This is where we write the facebook request id and facebook id of the person being invited.
When the person accepts the invitation, we are passed back that
person's facebook id and the facebook request id that we got when the
invitation was sent.

Knowing the facebook request id and the facebook id, we can query the
app_request table for the inviter's id.  With the inviter id and the
new person's person.id, we can write 2 records to the friends table
(bi-directional relationship)