Reworked the app request stuff.  Now, whenever you send an app request to someone, we write a record to the person table having that facebook id.
We turn this person into a friend of the current user and so the new
friend is visible on the Friends page.  Under the name is "pending" to
indicate that this person hasn't accepted the app request.

But the best part is: Because this person is now in the person table,
we can go immediately to his wish list and start adding stuff.  We
don't have to wait for him to accept the app request.

Also created app-CircleModule.js.  Starting to break stuff up into
modules so that I can define $rootScope (global) functions