Add Event modal:  Looks like this is working now.  It was buggy: You could open the modal once, then after than it wouldn't open.
All I could figure is that the button that opens the modal needs to
come after the markup containing the modal.

See modal_addevent.html...  There is a "plus sign" drop down button
that drops down the different kinds of events.  Selecting an event
causes the Add Event modal to appear.  Notice that the "plus sign"
button is at the very bottom of modal_addevent.html