Fixed name clash in app-EventCtrl.js:  There was a $scope.people that conflicted with the $rootScope.people from app.js: $rootScope.query().
$rootScope.query() is a function that finds all people with a given
name or part of name.  It used to return a list of people as
$rootScope.people.  But modal_addevent.html was using 'people' in 2
different ways.

The first way was the list of people that was being added to a circle.
The second way was the list of people that were returned when the user
went searching for people to add by name.

So I changed the name of the people that were returned as query
results from 'people' to 'peoplesearchresults'.

Had to update a lot of pages that are probably going to be going away.