Committing stuff that is still work-in-progress so I can discuss with Sudhanshu at Vercingetorix.
See app-EventCtrl.js -> NewEventCtrl and also EventsCtrl. The
routeChangeSuccess events seem to be firing when I navigate around the
app, but not when I hit F5/refresh in the browser.  

However, see app-CircleModule.js, line 56.  I'm listening for
routeChangeStart events here also.  But unlike in NewEventCtrl and
EventsCtrl, the function in app-CircleModule.js IS being called when I
do a browser refresh.

What's up with that?