bug fixes related to logging in... Used to be that when you logged in via FB, you would see the "Who Are You?" message at the top of the page. That's because we had the wrong ng-show clause in welcome-nofooter.html. We were checking to see if there was a 'user' or not.  Really what we needed to be looking for was 'users' - note plural.  And more specifically whether or not users.length was > 1
When there were multiple users found with a given email address, the
list we presented to the user didn't have the cool jquery mobile styling
applied.  The styling got lost as it often does due to the clash between
jqm and angular.  Created fn refreshWhoAreYouList() in user.js  This fn
reapplies the jqm styling.