Bug Fix (web):  In app-CircleModule:$rootScope.beginAddingFromAnotherEvent(), trying to be clever by removing the current circle from the list of circles that you could copy people from (because if you're adding people from another circle to the current circle, why would you want to display the current circle as one of your choices)
The problem is, there may not be a 'current' circle.  If you have just
come from 'my wish list' there won't be a 'current' circle.

This fix was very minor.  In $rootScope.beginAddingFromAnotherEvent(),
we create a copy of the user's circles, then we remove the 'current'
circle.  We weren't checking to see if there actually WAS a 'current'
circle.  We were just assuming there was one.

We updated the 'if' clause to first make sure that $rootScope.circle was
defined.