Enhancement (mobile): Implemented app update strategy.  There is now a table in the db named app_version.  There is a scala mapper class called AppVersion referenced in Boot.scala.  RestService now has a new url it responds to: /rest/version that queries the app_verion table via the AppVersion object and returns the value to the client.
On the client, there's a new function called $scope.version in
controllers.js (login.js) that calls the new url in RestService.
$scope.version is where we've hardcoded the version number of the
client/mobile app.  It should be the same value that you see in
config.xml in the <widget> tag.

If the version in $scope.version() is not the same as the version
returned from the server, then we set $scope.needupdate to true.

In home.html and welcome(-nofooter).html, we check the value of
$scope.needupdate and if it's true, we display a message to the user
that there is an update available.

Currently, we don't provide a link to the app store.  But we will as
soon as the app in is there.