In save, we now send an email to the person that was just added to whatever event.
I forgot to notify people when an event is first created.  The only
people that were getting emails were people that were added AFTER the
circle had been created - that was in RestService.insertParticipants.

I decided to put the email logic inside save so that wherever I needed
to call cp.save, I could be sure that emails would be sent.