web: Borrowing from the mobile client, I added the FacebookUser service. This service takes the first, last, email and fb id of the person logging in (via fb) and sends that info to the server.  The server figures out if this person is in the db yet or not.
Further down this file is a fn called initfbuser().  It's the function
that gets called right after we know we have a successful fb login.
This initfbuser() fn should get a lot simpler now that we have a
server-side utility to make sense of the name/email/fb id values.

In this version though, initfbuser() is still the really complicated
version that doesn't even use the new FacebookUser service