Changed the definition of the Email service by adding this parm ->
email:'@email'

This new parm makes the password recovery function on the website match
the password recovery function on the mobile app.  Up to now, the
website was using the "to" parameter to pass the email address.

The mobile app was using the "to" parameter as the name of the person
belonging to the email address - woops!